==> 01_compile_posts_data copy.ipynb <==
# Compile post data from json format 

Read in 166 ndjson files. Perform preliminary EDA on several to understand structure. Drop null values in text columns. Export a subset of a json to use as a sample dataset for preliminary modeling. Compile all posts into 1 dataframe. Export as a csv. 

import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt

pd.set_option('display.max_columns', None)

# df = pd.read_json('data/raw_ndjson/parler_data/parler_data000000000000.ndjson', lines=True)

df.shape

df.head()

df.columns

df.info()

# df.isna().sum()/len(df)*100

df['body'].unique

None of the rows have post data - ommiting first ndjson (00) out of 166

#### Post data file 01: data/raw_ndjson/parler_data/parler_data000000000001.ndjson

df1 = pd.read_json('data/raw_ndjson/parler_data/parler_data000000000001.ndjson', lines=True)

df1.to_csv('data/posts_01_raw.csv', index=False)

df1.shape

df1.head()

post_lengths = [len(str(each)) for each in df1['body']]

458803 rows just had URLs as body text so they are empty - drop all rows were body text is '' or ' '

df1['body'] = df1['body'].str.strip()

df1 = df1.loc[(df1['body'] != '')]

df1.loc[df1['body'] == '']

df1['body']

2 rows have null values in the body text column - drop 

df1['body'].isna().sum()

df1.dropna(subset=['body'], inplace=True)

df1.head(2)

sample_df = df1.sample(100_000, random_state=42)

sample_df.head()

# sample_df.to_csv('data/posts_01_raw_sample.csv', index=False)

## Import data files 

# df_2 = pd.read_json('data/raw_ndjson/parler_data/parler_data000000000002.ndjson', lines=True)
# df_2.to_csv('data/posts_02_raw.csv', index=False)
# df_3 = pd.read_json('data/raw_ndjson/parler_data/parler_data000000000003.ndjson', lines=True)
# df_3.to_csv('data/posts_03_raw.csv', index=False)
# df_164 = pd.read_json('data/raw_ndjson/parler_data/parler_data000000000164.ndjson', lines=True)
# df_164.to_csv('data/posts_164_raw.csv', index=False)
# df_165 = pd.read_json('data/raw_ndjson/parler_data/parler_data000000000165.ndjson', lines=True)
# df_165.to_csv('data/posts_165_raw.csv', index=False)
# df_166 = pd.read_json('data/raw_ndjson/parler_data/parler_data000000000166.ndjson', lines=True)
# df_166.to_csv('data/posts_166_raw.csv', index=False)

## Import CSV files and concatenate

csv_1 = pd.read_csv('data/raw_ndjson/posts_01_raw.csv')
csv_2 = pd.read_csv('data/raw_ndjson/posts_02_raw.csv')
csv_3 = pd.read_csv('data/raw_ndjson/posts_03_raw.csv')
csv_164 = pd.read_csv('data/raw_ndjson/posts_164_raw.csv')
csv_165 = pd.read_csv('data/raw_ndjson/posts_165_raw.csv')
csv_166 = pd.read_csv('data/raw_ndjson/posts_166_raw.csv')

np.random.seed(123)
json_list = list(np.random.randint(10, 150, 10))
json_list

for each in json_list:
    print(f"csv_{each} = pd.read_csv('data/raw_ndjson/posts_{each}_raw.csv')")
    # print(f'csv_{each}')

for each in json_list:
    print(f'csv_{each},')

list_csv = [csv_119,
csv_136,
csv_76,
csv_108,
csv_27,
csv_93,
csv_116,
csv_133,
csv_67,
csv_106,
]

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000119.ndjson', lines=True).to_csv('data/raw_ndjson/posts_119_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000136.ndjson', lines=True).to_csv('data/raw_ndjson/posts_136_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000076.ndjson', lines=True).to_csv('data/raw_ndjson/posts_76_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000108.ndjson', lines=True).to_csv('data/raw_ndjson/posts_108_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000027.ndjson', lines=True).to_csv('data/raw_ndjson/posts_27_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000093.ndjson', lines=True).to_csv('data/raw_ndjson/posts_93_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000116.ndjson', lines=True).to_csv('data/raw_ndjson/posts_116_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000133.ndjson', lines=True).to_csv('data/raw_ndjson/posts_133_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000067.ndjson', lines=True).to_csv('data/raw_ndjson/posts_67_raw.csv', index=False)

pd.read_json('data/raw_ndjson/parler_data/parler_data000000000106.ndjson', lines=True).to_csv('data/raw_ndjson/posts_106_raw.csv', index=False)

# csv_119 = pd.read_csv('data/raw_ndjson/posts_119_raw.csv')
# csv_136 = pd.read_csv('data/raw_ndjson/posts_136_raw.csv')
# csv_76 = pd.read_csv('data/raw_ndjson/posts_76_raw.csv')
# csv_108 = pd.read_csv('data/raw_ndjson/posts_108_raw.csv')
# csv_27 = pd.read_csv('data/raw_ndjson/posts_27_raw.csv')
# csv_93 = pd.read_csv('data/raw_ndjson/posts_93_raw.csv')
# csv_116 = pd.read_csv('data/raw_ndjson/posts_116_raw.csv')
# csv_133 = pd.read_csv('data/raw_ndjson/posts_133_raw.csv')
# csv_67 = pd.read_csv('data/raw_ndjson/posts_67_raw.csv')
# csv_106 = pd.read_csv('data/raw_ndjson/posts_106_raw.csv')

list_csv

posts_csv = [csv_1, csv_2, csv_3, csv_164, csv_165, csv_166]

[each.shape for each in posts_csv]

[each.shape for each in list_csv]

sum([each.shape[0] for each in posts_csv])

sum([each.shape[0] for each in list_csv])

all_posts_2 = pd.concat(list_csv)
all_posts_2.shape

The combined dataframe should have 6 million 573 thousand 256 rows 

all_posts = pd.concat(posts_csv)

all_posts.shape

all_posts.dropna(subset=['body'], inplace=True)

all_posts.shape

all_posts['body'] = all_posts['body'].str.strip()

all_posts = all_posts.loc[all_posts['body'] != '']

all_posts.loc[all_posts['body'] == '  ']

all_posts.shape

all_posts['body']

all_posts.reset_index(inplace=True)

all_posts['creator'].unique().shape

all_posts_2.shape

all_posts_2.dropna(subset=['body'], inplace=True)

all_posts_2.shape

all_posts_2['body'] = all_posts_2['body'].str.strip()

all_posts_2 = all_posts_2.loc[all_posts_2['body'] != '']

all_posts_2.loc[all_posts_2['body'] == ' ']

all_posts_2.shape

all_posts_2['body']

all_posts_2.reset_index(inplace=True)

all_posts_2['creator'].unique().shape

combined_df = pd.concat([all_posts, all_posts_2], ignore_index=True).reset_index()

combined_df.shape

combined_df.head(1)

combined_df.to_csv('data/raw_ndjson/all_posts.csv', index=False)

unique_creators = list(set(combined_df['creator']))

len(unique_creators)//2

split_610395_creators = unique_creators[:610395]

half_df = combined_df.loc[combined_df['creator'] == split_610395_creators]
