==> 03_clean_compiled_datasets copy.ipynb <==
# Clean users and posts compiled datasets 

For both the users and posts datasets:
- Pare down features
- Drop nulls and deal with missing data
- Remove usernames and html artifacts 
- Join users to post dataset
- Export cleaned data as 2 separate csvs: modeling (for train and test) and validation 

import numpy as np
import pandas as pd
import re

pd.set_option('display.max_columns', 100)
pd.set_option('display.max_colwidth', 500)

# posts = pd.read_csv('data/raw_ndjson/all_posts.csv')
# posts.columns = posts.columns.str.lower()
# # posts.shape

posts.head(3)

## Drop columns

df = posts[['body', 'createdatformatted', 'creator', 'id', 'username', 'reposts', 'score']]

df.head(2)

df.shape

## Clean up text with regex 

# remove usernames
df['body'] = df['body'].map(lambda x: re.sub("\@[a-zA-Z0-9]*", ' ', str(x)))

# remove new lines characters and html artifacts
df['body'] = df['body'].map(lambda x: re.sub("\n|\r|&amp;#x200B;|&amp;", ' ', str(x)))

# remove non ascii characters; ommits emojis, ommits arabic 
# regex from https://stackoverflow.com/questions/61721443/how-to-remove-or-filter-non-english-chinese-korean-japanese-arabic-strings
df['body'] = df['body'].map(lambda x: re.sub("[^\u0000-\u05C0\u2100-\u214F]+", ' ', str(x).strip()))

df.shape

regexed = df

# drop empty rows (had either emojis or arabic)
df = df.loc[df['body'] != ' ']
df = df.loc[df['body'] != '']

df.loc[df['body'] == ' ']

df.shape

# arabic_examples

(df.isna().sum()/len(df) * 100).sort_values()

# drop rows with null usernames with null values or blanks 
df = df.loc[df['username'].isna() == False]

df.loc[df['username'] == ' ']

df.shape

cleaned_df = df

## Create transposed table with creator, body text, timestamp 

users_posts = df[['body', 'username']]

users_posts.shape

users_posts['username'].value_counts().sort_values().tail(25)

top_posters = ['TTexasRepublic', 'GaryADePietro', 'robertleon', 'moon52', 'Oldmanrant', 'ThomasFox', 'Gayle7753', 'chucknellis', 'doutingthomas1', 'Mikewin', 'KimberlyHumphreys', 'Klonokid', 'LibertyElaine', 'Blackshoe', 'RadianceLux', 'EliseRhodes', 'jenniev101', 'TeamTrump', 'WashTimesOpEd', 'JoePags', 'ronpaul', 'John']

for each in top_posters:
    print(each)
    print(users_posts.loc[users_posts['username'] == each].value_counts().sort_values().tail(2))
    print()

## Remove posts from welcome bots
- Welcome to Parler
- #parlerconcierge 
- #ParlerConcierge 

Team trump is a bot but not associated with Parler - removing the welcome posts but not the account 

concierge_bots = ['John', 'ronpaul', 'JoePags', 'WashTimesOpEd', 'LibertyElaine', 'blackshoe', 'radiancelux', 'eliserhodes', 'jenniev101', 'KimberlyHumphreys']

len(df)

df = df.loc[~df['username'].isin(concierge_bots)]

without_bots = len(df)
without_bots

df = df[df['body'].str.contains('Welcome to Parler') == False]

without_welcometoparler = len(df)
without_welcometoparler

df = df[df['body'].str.contains('#parlerconcierge') == False]

df = df[df['body'].str.contains('#ParlerConcierge') == False]

without_parlerconcierge = len(df)
without_parlerconcierge

## Combine tweets for each user 

users_posts = df[['body', 'username']]

# code for combing rows https://stackoverflow.com/questions/36392735/how-to-combine-multiple-rows-into-a-single-row-with-pandas
combined_posts = users_posts.groupby('username')['body'].apply(' '.join).reset_index()

combined_posts.head(10)

# arabic_examples = combined_posts.head(10)

combined_posts.sample(15)

combined_posts['username'].value_counts()

## Export cleaned posts and transposed posts csvs

df.to_csv('data/cleaned_posts.csv', index=False)

combined_posts.to_csv('data/posts_by_user.csv', index=False)

## Create smaller sample of cleaned_posts

import_cleaned = pd.read_csv('data/cleaned_posts.csv')

sample = import_cleaned.sample(frac=.1, random_state=42)

sample.info()

users_posts_sample = sample[['body', 'username']]

users_posts_sample.head()

users_posts_sample.groupby('username')['body'].describe()

# code for combing rows https://stackoverflow.com/questions/36392735/how-to-combine-multiple-rows-into-a-single-row-with-pandas
combined_posts_sample = users_posts_sample.groupby('username')['body'].apply(' '.join).reset_index()

len(combined_posts_sample['username'].unique())

# combined_posts_sample.to_csv('data/posts_by_users_sample.csv', index=False)

# sample.to_csv('data/cleaned_posts_10sample.csv', index=False)
