==> 05_text_preprocessing_posts copy.ipynb <==
# Text preprocessing

Steps 
- Remove usernames and html artifacts 
- Split into train, test, and validation 
- Create word count matrix and remove stop words with CountVectorizer (demonstration, will repeat with the same random state for each step)
- Export cleaned test, cleaned train, cleaned validation, and countvectorized 

import numpy as np 
import pandas as pd 
import re 

from sklearn.feature_extraction.text import CountVectorizer

pd.set_option('display.max_columns', 100)
pd.set_option('display.max_colwidth', 500)

## Load text data

df = pd.read_csv('../data/posts_01_raw_sample.csv')

df.head(2)

df = df[['body']]

# need to dropna in importing step with entire df - REMOVE
df = df.dropna().reset_index(drop=True)

df.tail(20)

df.tail(15)

df.shape

df.isna().sum()

## CountVectorize and remove stop words

default_words = list(CountVectorizer(stop_words='english').get_stop_words())

# common contractions missed by CountVectorizer 
contractions = ['ve', 're']

# add custom lists to default sklearn 'english' stopwords 
custom_stopwords = default_words + contractions

X = df['body']

cv = CountVectorizer(min_df = 10, 
                     max_df = .5,
                     stop_words = custom_stopwords, 
                     ngram_range = (1, 3),
                     strip_accents='ascii', 
                )
cv.fit(X)

cv_text = cv.transform(X)
cv_df = pd.DataFrame(cv_text.todense(), columns=cv.get_feature_names_out())
cv_df.head(2)

cv_df.columns[3050:3100]

[each for each in cv_df.columns if 'parler' in each]

cv_df.columns[5000:5050]

cv_df.shape

(58016, 5220)
(58016, 51194)


#### After initial CountVectorizer run decided to go back to original compiled raw dataset and:
- remove all posts with the word "parler" (see parler_bot for examples of texts using the words parler)
- 

parler_bot = ['aboard parler', 'aboard parler just', 'best parler', 'best parler experience', 'best rob parlerconcierge', 'check hashtag parlerconcierge', 'check parler', 'check parler youtube', 'choosing parler', 'choosing parler free', 'congratulations choosing parler', 'ensure best parler', 'espanol tambien parlerconcierge', 'glad parler', 'glad parler free', 'glad parler truly', 'hashtag parlerconcierge', 'hashtag parlerconcierge learn', 'joined parler', 'joined parler looking', 'just joined parler', 'know questions parlerconcierge', 'like parler', 'like parler accepts', 'looking parler', 'looking parler tips', 'newuser truefreespeech parlerconcierge', 'parler', 'parler accepts', 'parler accepts right', 'parler comments', 'parler comments leave', 'parler experience', 'parler experience puedo', 'parler free', 'parler free speech', 'parler help', 'parler help make', 'parler hope', 'parler hope enjoy', 'parler just', 'parler just wanted', 'parler looking', 'parler looking forward', 'parler parlerusa', 'parler pro', 'parler pro tips', 'parler step', 'parler step taking', 'parler tips', 'parler tips tos', 'parler truly', 'parler truly trump', 'parler youtube', 'parler youtube page', 'parler101', 'parler101 channel', 'parler101 channel good', 'parlerconcierge', 'parlerconcierge learn', 'parlerconcierge learn ropes', 'parlerksa', 'parleruk', 'parlerusa', 'people like parler', 'people looking parler', 'places parler101', 'places parler101 channel', 'questions parlerconcierge', 'rob parlerconcierge', 'share parler', 'share parler pro', 'tambien parlerconcierge', 'tos check parler', 'truefreespeech parlerconcierge', 'videos places parler101', 'wanted share parler', 'welcome aboard parler', 'welcome glad parler', 'welcome parler', 'welcome parler comments', 'welcome parler help', 'welcome parler hope', 'welcome parler step']

## Check top n-grams and refine stop words 

# print(default_words)

# remove numbers? 

## Split into validation vs modeling (modeling will be further split into train and test)

## Export cleaned validation and modeling text datasets

# df.to_csv('../data/posts_01_cleaned_sample.csv', index=False)
