==> 07_LDA_by_post copy.ipynb <==
# LDA by individual post 

Each vector is a distinct post. Predictions would be performed on single new posts to determine individual probability of belonging to a cluster. 

Code adapted from 
- General Assembly Lesson by Caroline Schmitt: Intro to Latent Dirichlet Allocation. 
- Tutorial from Machine Learning Plus by Selva Prabhakaran https://www.machinelearningplus.com/nlp/topic-modeling-python-sklearn-examples/

import numpy as np 
import pandas as pd 
import matplotlib.pyplot as plt 

from sklearn.feature_extraction.text import CountVectorizer 
from sklearn.decomposition import LatentDirichletAllocation 
from sklearn.model_selection import train_test_split, GridSearchCV

np.random.seed(42)

## Import preprocessed data

df = pd.read_csv('data/posts_01_cleaned_sample.csv')

df.head()

# need to dropna in importing step with entire df - REMOVE
df = df.dropna().reset_index(drop=True)

df.head()

## Train test split

X = df['body']

X_train, X_test = train_test_split(X, test_size=0.1, random_state=42)

## Count Vectorize

default_words = list(CountVectorizer(stop_words='english').get_stop_words())
custom_stopwords = default_words + ['ve', 're']

cv = CountVectorizer(min_df = 3, 
                     max_df = .5, 
                     stop_words = custom_stopwords, 
                     strip_accents='ascii', 
                )
cv.fit(X_train)

cv_train = cv.transform(X_train)
cv_test = cv.transform(X_test)
cv_df = pd.DataFrame(cv_train.todense(), columns=cv.get_feature_names_out())
cv_df.head(2)

## Fit LDA model

lda = LatentDirichletAllocation(n_components=5, 
                                random_state=42
                               )
lda.fit(cv_train)

len(lda.components_)

for each in lda.components_:
    print(len(each))

features = cv.get_feature_names_out()

def display_topics(model, feature_names, num_words):
    for ix, topic in enumerate(model.components_):
        print(f'Topic {ix}:')
        words = [feature_names[i] for i in topic.argsort()[:-num_words - 1:-1]]
        print('\n'.join(words))
        print()


display_topics(lda, features, 10)

lda_8 = LatentDirichletAllocation(n_components=8, 
                                random_state=42
                               )
lda_8.fit(cv_train)

len(lda.components_)

display_topics(lda, features, 15)

lda_8 = LatentDirichletAllocation(n_components=25, 
                                random_state=42
                               )
lda_8.fit(cv_train)

len(lda.components_)

display_topics(lda, features, 10)

## Examine words in topics
